requirements.txt
tensorflow-gpu==1.11.0
imageio==2.5.0
matplotlib==3.0.3
numpy==1.15.2
Pillow==5.2.0
scikit-image==0.15
scikit-learn==0.19.2
scipy==1.1.0
sklearn
fasteners
fjcommon==0.1.69
lazyme==0.0.22
tqdm==4.32.1

